Reset inet_db timer even if file is up to date
Prior to this patch, if 5 seconds passed and
the resolv_conf or hosts_file were up to date,
we would continue to check for the file update
on every inet_res or inet_hosts operation.

This PR makes sure that the timer is updated
even if the file is up to date, so we check
again only in the next 5 seconds.